requirements.txt
enhanced-dpy
python-dotenv
uvloop